# docs/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'Foodgram'
  version: ''
paths:
...
  /api/ingredients/:
    get:
      operationId: Список ингредиентов
      description: 'Список ингредиентов с возможностью поиска по имени.'
      parameters:
        - name: name
          required: false
          in: query
          description: Поиск по частичному вхождению в начале названия ингредиента.
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'
          description: ''
      tags:
        - Ингредиенты
...
components:
  schemas:
...
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 128
          example: 'Капуста'
        measurement_unit:
          type: string
          maxLength: 64
          example: 'кг'
      required:
        - name
        - measurement_unit
...
  securitySchemes:
    Token:
      description: 'Авторизация по токену. <br>'
      Все запросы от имени пользователя должны выполняться с заголовком "Authorization: Token TOKENVALUE"'
      type: http
      scheme: token